Pull in latest genenetwork3 when running genenetwork2 tests.
So far, whenever genenetwork2 code needed the latest updates from
genenetwork3, the guix-bioinformatics channel had to be updated and
the development container had to be rebuilt. No more. From now on, the
latest genenetwork3 code will be pulled in at all times. The
development container need only be rebuilt when the dependencies of
genenetwork2 or genenetwork3 change. We may revisit this decision once
genenetwork2 and genenetwork3 are better decoupled and we have
semantically versioned releases.

* genenetwork-development.scm (genenetwork2-tests): Pull in latest
genenetwork3 when running genenetwork2 tests.