Document getting a shell into container
For some tasks (e.g. debugging the xapian index build), we need to
manually queue the job, which means we need a shell into the
container.

Getting a shell using the default command displayed actually fails
with:

    nsenter: failed to execute /bin/bash: No such file or directory

so we need to provide the correct path to bash, and optionally,
initialise the shell to setup the correct paths.